+++ packages/ketchup/src/components/kup-data-table/kup-data-table.scss
@@ -4,7 +4,7 @@
 @import 'modules/kup-data-table-fixed-rows';
 @import 'modules/kup-data-table-drag-drop';
 @import 'modules/kup-data-table-classes';
-@import '../../f-components/image/f-image.scss';
+@import '../../utils/components/f-image/f-image.scss';
 
 :host {
   display: block;
